+++ packages/wonder-blocks-pill/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @khanacademy/wonder-blocks-pill
 
+## 2.0.0
+
+### Major Changes
+
+-   441380b0: Pill now has small, medium, and large sizes. Medium is the new default size. This change was made because the small size is a bit too small as per
+    accessibility recommendations, but is still necessary in some cases.
+    The medium size is the recommended "smaller" size. The large pill is the same.
+
 ## 1.1.0
 
 ### Minor Changes